feat(MediaPlayerBar): add CurrentEpisode component
